requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.0
python-dotenv==0.19.2
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
